Re-design: move db connect to it's own file.
